requirements.txt
openai==1.12.0
requests==2.31.0
groq>=0.4.0
jira==3.5.2
PyGithub==1.59.1
boto3==1.34.34
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiohttp==3.9.1
aiofiles==23.2.0
yt-dlp==2024.1.6
beautifulsoup4==4.12.2
lxml==4.9.4
python-multipart==0.0.6
